+++ app/global-error.tsx
@@ -1,7 +1,7 @@
 'use client' // Error components must be Client Components
 
 import { ErrorPageLayout } from '@/components/layout/ErrorPageLayout'
-import { ID } from '@/utils/constants/ids/ids'
+import { ID } from '@/utils/constants/ids/elementIds'
 import { ErrorProps } from '@/utils/shared/componentProps'
 
 export default function GlobalError({ error, reset }: ErrorProps) {
