+++ app/work-experience/smartsupp-dashboard/page.tsx
@@ -3,7 +3,7 @@ import PageNavigation from '@/components/pageNavigation/PageNavigation'
 import { projectsWorkReact } from '@/data/projects/work/projectsOverview/workReact'
 import { sections } from '@/data/projects/work/smartsuppDashboard'
 import { TEXT } from '@/localization/texts_en'
-import { ID } from '@/utils/constants/ids/ids'
+import { ID } from '@/utils/constants/ids/elementIds'
 import { PAGES_URL, URL_WORK_EXPERIENCE } from '@/utils/constants/pageUrls'
 import { DATA_TEST_IDS } from '@/utils/dataTestIds'
 import { BreadCrumbsType } from '@/utils/types'
