+++ components/homepage/Hero.tsx
@@ -1,4 +1,4 @@
-import { ID } from '@/utils/constants/ids/ids'
+import { ID } from '@/utils/constants/ids/elementIds'
 import HeroHeading from './HeroHeading'
 import HeroMainLinks from './HeroMainLinks'
 import HeroParagraph from './HeroParagraph'
